+++ LookupAnything/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed NPC lookups showing gift tastes for flavored variants in 1.32.3.
+* Fixed tree lookups incorrectly stating that a fertilized tree won't grow in winter.
 * Fixed error when trying to lookup on the 'help wanted' screen.
 * Fixed untranslated text in child lookup.
 * Improved translations. Thanks to Kareolis (updated Russian)!